// packages/theme-generator/src/generatedFlavors.ts
export interface FlavorScheme {
  scheme: string;
  author: string;
  slug?: string;
  base00: string;
  base01: string;
  base02: string;
  base03: string;
  base04: string;
  base05: string;
  base06: string;
  base07: string;
  base08: string;
  base09: string;
  base0A: string;
  base0B: string;
  base0C: string;
  base0D: string;
  base0E: string;
  base0F: string;
}
// ...
export const FLAVOR_NAMES: string[] = [
  "catppuccin-latte",
  "catppuccin-mocha",
  "dracula",
  "github",
  "gruvbox-dark",
  "gruvbox-light",
  "nord",
  "rose-pine-dawn",
  "solarized-light",
  "tokyo-night-dark"
];